+++ bitcoin/src/bip158.rs
@@ -561,7 +561,6 @@ mod test {
     use super::*;
     use crate::consensus::encode::deserialize;
     use crate::hash_types::BlockHash;
-    use crate::hashes::hex::FromHex;
     use crate::ScriptBuf;
     use crate::internal_macros::hex;
