+++ bitcoin/src/crypto/ecdsa.rs
@@ -8,11 +8,10 @@
 use core::str::FromStr;
 use core::{fmt, iter};
 
-use internals::write_err;
+use internals::{hex, write_err};
 use secp256k1;
 
 use crate::prelude::*;
-use crate::hashes::hex::{self, FromHex};
 use crate::sighash::{EcdsaSighashType, NonStandardSighashType};
 
 /// An ECDSA signature with the corresponding hash type.
