+++ bitcoin/src/network/address.rs
@@ -307,9 +307,8 @@ mod test {
     use core::str::FromStr;
     use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr};
 
-    use super::{AddrV2, AddrV2Message, Address};
+    use super::*;
     use crate::consensus::encode::{deserialize, serialize};
-    use crate::hashes::hex::FromHex;
     use crate::internal_macros::hex;
     use crate::network::constants::ServiceFlags;
