+++ bitcoin/src/psbt/serialize.rs
@@ -437,7 +437,6 @@ fn key_source_len(key_source: &KeySource) -> usize {
 mod tests {
     use core::convert::TryFrom;
 
-    use crate::hashes::hex::FromHex;
     use super::*;
 
     // Composes tree matching a given depth map, filled with dumb script leafs,
