+++ bitcoin/src/serde_utils.rs
@@ -27,7 +27,6 @@ pub mod btreemap_byte_values {
 
     use serde;
 
-    use crate::hashes::hex::FromHex;
     use crate::prelude::*;
 
     pub fn serialize<S, T>(v: &BTreeMap<T, Vec<u8>>, s: S) -> Result<S::Ok, S::Error>
@@ -251,7 +250,7 @@ pub mod hex_bytes {
 
     use serde;
 
-    use crate::hashes::hex::FromHex;
+    use crate::prelude::*;
 
     pub fn serialize<T, S>(bytes: &T, s: S) -> Result<S::Ok, S::Error>
     where
